requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.43.0
streamlit>=1.28.0
plotly>=5.18.0
yfinance>=0.2.31
fredapi>=0.5.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
joblib>=1.3.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
ipykernel>=6.26.0
